move node.ts to backend/index.ts
